Change logic to select a value from a csv column to check in order
The function merge:entry accepts a sequence of strings to look for column names that match any of them.

The previous logic was selecting the first column with a matching name.
The new logic selects the column with the first name in the sequence provided: this makes the list names work as priorities.

Also, empty values will be skipped.

This is applied in the Results template, to look for "Period Start/End Date", "Start/End Date" in that order (export format of Toladata).

merge:entry with arity 2 now calls merge:entry with arity 3

Fixes #110