Add "entryExists" function to allow skipping non-present columns
While adding the "ascending" attribute, it emerged that not all client templates contain that column yet, and a default "false" was added.

In a broader sense, being able to add optional columns to some but not all client templates is useful. The entryExists function allows testing for the presence of a column.