Fix publishing process, remove dataworkbench call
Replaced call to dataworkbench with copy operations of only IATI XML files.
(We don't use the static version of the workbenches, generating those is no longer needed.)

Added property dwb.organisation to iati-workbench.properties (added for all client workbenches).
Create a separate issue to make this more robust, but changing this to copying the files to Google Storage should be straightforward.