Remove lib-workspace, update README
Since everything is presented in the container, an external workspace helper build.xml is no longer needed.

The extension points are moved into the S2I build targets, but cannot be extended at this moment. Kept for consideration or future removal.

README slightly updated to indicate the focus of this version.