fix: improve handling of multilingual narratives
The xml:lang attribute will be generated for narratives defined in Projects.
The first xml:lang attribute in the generated IATI activity xml sequence will
be used as default language (this should be the first occurrence of an activity
in the Projects sheet).
Narrative fields will add an xml:lang attribute for non-default languages.
For elements title and description, the narratives are collected under that
element.