debt: simplify test setup for default templates
With run-as=external and file as parameter, it is possible
to run all tests for default templates from a single master xspec file.
This also results in a single report for all those tests.

Also fix a few references in included xspec files.
(Since the stylesheet to test is determined in the master xspec file,
the value does not really matter, but now points to the proper xslt)