docs: create docs for multiple versions
By default, the Github checkout action only checks out the commit that triggered the workflow.
As a result, using the local repository as content source will only produce a single version for it.

Update to v3.
Checkout everything: it's what Antora would do when provided a full URL for the source.